boards: Ti60F225: Fix the timebase-frequency to 50Mhz
timebase-frequency is referring to the frequency at which the RISC-V
rdtime CSR count up. In the soc, that CSR read the value from the
CLINT peripheral, which is running in the peripheral clock domain.

The Ti60F225 board configure with 250Mhz of cpu frequency
and 50Mhz of peripheral clock frequency. When the timebase-frequency
is set to 250Mhz, the timer become 5 times slower than real time clock.
For example, 1 second of board time equal to 5 seconds in real time.
Thus, set the timebase-frequency to 50Mhz fix this issue.

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>