kernel_modules/gpio: add GPIO consumer driver
The GPIO consumer driver will drive the GPIO pin
to input and output. By default, only 4 GPIO pin
get enable on both Ti60 and T120 board. When the
driver is loaded, the GPIO pin can be configure
in /sys/class/efx_gpios for direction and value
of sysfs attribute. Direction 1 for output and 0
for input.

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>