device_tree_generator.py: add function to get supported peripheral
The generation of device tree node should be dynamic as Sapphire SoC
can be configured as needed. For example, user might remove I2C peripheral
from the Sapphire SoC during the SoC generation in Efinity IPM. Thus, the
device tree node should be generated based on the supported peripheral in
the SoC configuration.

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>